+++ packages/guard/src/index.ts
@@ -1,2 +1,1 @@
-export { Guard } from './guard.gen'
 export * from './signer'