+++ packages/waas/src/email.ts
@@ -6,8 +6,15 @@ import {
   SignUpCommand,
   UserLambdaValidationException
 } from '@aws-sdk/client-cognito-identity-provider'
+import { Identity } from './auth'
 
-import { IdTokenIdentity } from './auth'
+function getRandomString(len: number) {
+  const randomValues = new Uint8Array(len)
+  window.crypto.getRandomValues(randomValues)
+  return Array.from(randomValues)
+    .map(nr => nr.toString(16).padStart(2, '0'))
+    .join('')
+}
 
 export class EmailAuth {
   private cognitoMemo: CognitoIdentityProviderClient
@@ -28,7 +35,6 @@ export class EmailAuth {
   }
 
   private signUp(email: string) {
-    email = email.toLowerCase().trim()
     return this.cognito().send(
       new SignUpCommand({
         ClientId: this.clientId,
@@ -40,7 +46,6 @@ export class EmailAuth {
   }
 
   private signIn(email: string) {
-    email = email.toLowerCase().trim()
     return this.cognito().send(
       new InitiateAuthCommand({
         AuthFlow: 'CUSTOM_AUTH',
@@ -54,7 +59,6 @@ export class EmailAuth {
 
   public async initiateAuth({ email }: { email: string }): Promise<{ email: string; instance: string }> {
     let res: InitiateAuthCommandOutput
-    email = email.toLowerCase().trim()
 
     try {
       // Try sign in directly first
@@ -91,9 +95,7 @@ export class EmailAuth {
     email: string
     answer: string
     sessionHash: string
-  }): Promise<IdTokenIdentity> {
-    email = email.toLowerCase().trim()
-
+  }): Promise<Identity> {
     const res = await this.cognito().send(
       new RespondToAuthChallengeCommand({
         ClientId: this.clientId,
@@ -109,26 +111,5 @@ export class EmailAuth {
     }
 
     return { idToken: res.AuthenticationResult.IdToken }
-  }
-}
-
-function getRandomString(len: number) {
-  return Array.from(getRandomValues(len))
-    .map(nr => nr.toString(16).padStart(2, '0'))
-    .join('')
-}
-
-function getRandomValues(len: number) {
-  const randomValues = new Uint8Array(len)
-  if (typeof window === 'object' && typeof window.crypto === 'object') {
-    return window.crypto.getRandomValues(randomValues)
-  } else {
-    console.warn('window.crypto.getRandomValues is not available. Falling back to less secure Math.random().')
-    const randomValues = new Uint8Array(len)
-    for (let i = 0; i < len; i++) {
-      const randomInteger = Math.floor(Math.random() * 256)
-      randomValues[i] = randomInteger
-    }
-    return randomValues
   }
 }