+++ packages/waas/src/intents/base.ts
@@ -1,5 +1,4 @@
-import { VERSION as PACKAGE_VERSION } from '@0xsequence/core'
-import { Intent as RawIntent, IntentName } from '../clients/intent.gen'
+import { Intent as RawIntent } from '../clients/intent.gen'
 import { useLifespan } from './utils'
 import { ethers } from 'ethers'
 import { canonicalize } from 'json-canonicalize'
@@ -8,10 +7,9 @@ import { Session } from '../session'
 export type Intent<T> = Omit<RawIntent, 'data' | 'signatures'> & { data: T }
 export type SignedIntent<T> = Omit<RawIntent, 'data'> & { data: T }
 
-const INTENTS_VERSION = 1
-const VERSION = `${INTENTS_VERSION} (Web ${PACKAGE_VERSION})`
+const VERSION = '0.0.0'
 
-export function makeIntent<T>(name: IntentName, lifespan: number, data: T): Intent<T> {
+export function makeIntent<T>(name: string, lifespan: number, data: T): Intent<T> {
   const issuedAt = Math.floor(Date.now() / 1000)
   const expiresAt = issuedAt + lifespan
   return {