+++ packages/waas/src/session/index.ts
@@ -1,7 +1,5 @@
 import { newSECP256K1SessionFromSessionId, newSECP256K1Session } from './secp256k1'
 import { newSECP256R1SessionFromSessionId, newSECP256R1Session } from './secp256r1'
-import { SubtleCryptoBackend } from '../subtle-crypto'
-import { SecureStoreBackend } from '../secure-store'
 
 export type Session = {
   sessionId(): Promise<string>
@@ -9,32 +7,19 @@ export type Session = {
   clear(): void
 }
 
-export async function newSessionFromSessionId(
-  sessionId: string,
-  cryptoBackend: SubtleCryptoBackend | null,
-  secureStoreBackend: SecureStoreBackend | null
-): Promise<Session> {
-  if (!secureStoreBackend) {
-    throw new Error('No secure store available')
-  }
-  if (cryptoBackend) {
-    return newSECP256R1SessionFromSessionId(sessionId, cryptoBackend, secureStoreBackend)
+export async function newSessionFromSessionId(sessionId: string): Promise<Session> {
+  if (window.crypto !== undefined) {
+    return newSECP256R1SessionFromSessionId(sessionId)
   } else {
-    return newSECP256K1SessionFromSessionId(sessionId, secureStoreBackend)
+    return newSECP256K1SessionFromSessionId(sessionId)
   }
 }
 
-export async function newSession(
-  cryptoBackend: SubtleCryptoBackend | null,
-  secureStoreBackend: SecureStoreBackend | null
-): Promise<Session> {
-  if (!secureStoreBackend) {
-    throw new Error('No secure store available')
-  }
-  if (cryptoBackend) {
-    return newSECP256R1Session(cryptoBackend, secureStoreBackend)
+export async function newSession(): Promise<Session> {
+  if (window.crypto !== undefined) {
+    return newSECP256R1Session()
   } else {
-    return newSECP256K1Session(secureStoreBackend)
+    return newSECP256K1Session()
   }
 }