+++ packages/waas/src/session/secp256k1.ts
@@ -1,19 +1,16 @@
 import { ethers } from 'ethers'
-import { SecureStoreBackend } from '../secure-store'
+import { openDB } from 'idb'
 import { Session } from './index'
 
 const idbName = 'seq-waas-session-p256k1'
 const idbStoreName = 'seq-waas-session'
 
-export async function newSECP256K1SessionFromSessionId(
-  sessionId: string,
-  secureStoreBackend: SecureStoreBackend
-): Promise<Session> {
-  const privateKey = await secureStoreBackend.get(idbName, idbStoreName, sessionId)
+export async function newSECP256K1SessionFromSessionId(sessionId: string): Promise<Session> {
+  const db = await openDB(idbName)
 
-  if (!privateKey) {
-    throw new Error('No private key found')
-  }
+  const tx = db.transaction(idbStoreName, 'readonly')
+  const privateKey = await db.get(idbStoreName, sessionId)
+  await tx.done
 
   const wallet = new ethers.Wallet(privateKey)
 
@@ -24,25 +21,33 @@ export async function newSECP256K1SessionFromSessionId(
     sign(message: string | Uint8Array): Promise<string> {
       return wallet.signMessage(message)
     },
-    clear: async () => {
-      await secureStoreBackend.delete(idbName, idbStoreName, sessionId)
+    clear(): void {
+      db.delete(idbStoreName, sessionId)
     }
   } as Session
 }
 
-export async function newSECP256K1SessionFromPrivateKey(
-  privateKey: string,
-  secureStoreBackend: SecureStoreBackend
-): Promise<Session> {
+export async function newSECP256K1SessionFromPrivateKey(privateKey: string): Promise<Session> {
   const wallet = new ethers.Wallet(privateKey)
+
+  const db = await openDB(idbName, 1, {
+    upgrade(db) {
+      db.createObjectStore(idbStoreName)
+    }
+  })
+
   const sessionId = await wallet.getAddress()
 
-  await secureStoreBackend.set(idbName, idbStoreName, sessionId, privateKey)
+  const tx = db.transaction(idbStoreName, 'readwrite')
+  await db.put(idbStoreName, privateKey, sessionId)
+  await tx.done
 
-  return newSECP256K1SessionFromSessionId(sessionId, secureStoreBackend)
+  db.close()
+
+  return newSECP256K1SessionFromSessionId(sessionId)
 }
 
-export async function newSECP256K1Session(secureStoreBackend: SecureStoreBackend): Promise<Session> {
+export async function newSECP256K1Session(): Promise<Session> {
   const wallet = ethers.Wallet.createRandom()
-  return newSECP256K1SessionFromPrivateKey(wallet.privateKey, secureStoreBackend)
+  return newSECP256K1SessionFromPrivateKey(wallet.privateKey)
 }