+++ packages/0xsequence/CHANGELOG.md
@@ -1,5 +1,32 @@
 # 0xsequence
 
+## 2.0.0
+
+### Major Changes
+
+- changeset
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsequence/abi@2.0.0
+  - @0xsequence/account@2.0.0
+  - @0xsequence/api@2.0.0
+  - @0xsequence/auth@2.0.0
+  - @0xsequence/core@2.0.0
+  - @0xsequence/guard@2.0.0
+  - @0xsequence/indexer@2.0.0
+  - @0xsequence/metadata@2.0.0
+  - @0xsequence/migration@2.0.0
+  - @0xsequence/multicall@2.0.0
+  - @0xsequence/network@2.0.0
+  - @0xsequence/provider@2.0.0
+  - @0xsequence/relayer@2.0.0
+  - @0xsequence/sessions@2.0.0
+  - @0xsequence/signhub@2.0.0
+  - @0xsequence/utils@2.0.0
+  - @0xsequence/wallet@2.0.0
+
 ## 1.10.14
 
 ### Patch Changes