+++ packages/deployer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @0xsequence/deployer
 
+## 2.0.0
+
+### Major Changes
+
+- changeset
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsequence/utils@2.0.0
+
 ## 1.10.14
 
 ### Patch Changes