+++ packages/waas/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @0xsequence/waas
 
+## 2.0.0
+
+### Major Changes
+
+- changeset
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsequence/core@2.0.0
+  - @0xsequence/network@2.0.0
+
 ## 1.10.14
 
 ### Patch Changes